tyny refactoring
