+++ options/index.js
@@ -1,0 +1,5 @@
+import * as opt from "./options";
+
+export const options = {
+  ...opt
+};
