+++ options/options.js
@@ -1,0 +1,66 @@
+export const categories = [
+  "business",
+  "entertainment",
+  "general",
+  "health",
+  "science",
+  "sports",
+  "technology"
+];
+
+export const country = [
+  "ae",
+  "ar",
+  "at",
+  "au",
+  "be",
+  "bg",
+  "br",
+  "ca",
+  "ch",
+  "cn",
+  "co",
+  "cu",
+  "cz",
+  "de",
+  "eg",
+  "fr",
+  "gb",
+  "gr",
+  "hk",
+  "hu",
+  "id",
+  "ie",
+  "il",
+  "in",
+  "it",
+  "jp",
+  "kr",
+  "lt",
+  "lv",
+  "ma",
+  "mx",
+  "my",
+  "ng",
+  "nl",
+  "no",
+  "nz",
+  "ph",
+  "pl",
+  "pt",
+  "ro",
+  "rs",
+  "ru",
+  "sa",
+  "se",
+  "sg",
+  "si",
+  "sk",
+  "th",
+  "tr",
+  "tw",
+  "ua",
+  "us",
+  "ve",
+  "za"
+];
